+++ src/core/App/theme.js
@@ -8,7 +8,7 @@ export const theme = {
         waterloo: '#7E839A',
         black: '#000000',
     },
-    breakpoint: {
+    breakpoints: {
         mobileNormal: '550px',
         mobileMax: '767px',
         tablet: '1000px',
